@@ -11,6 +11,10 @@
     "./react/*": "./react/*",
     "./*": "./elements/*"
   },
+  "imports": {
+    "#uxdot/*": "./*",
+    "#11ty-plugins/*": "./docs/_plugins/*"
+  },
   "contributors": [
     "Adam Johnson <adjohnso@redhat.com>",
     "Benny Powers <bennyp@redhat.com>",
@@ -289,18 +293,18 @@
   },
   "dependencies": {
     "@11ty/eleventy-upgrade-help": "^3.0.1",
-    "@lit/context": "^1.1.2",
+    "@lit/context": "^1.1.3",
     "@patternfly/pfe-core": "^4.0.4",
     "@rhds/icons": "^1.1.2",
     "@rhds/tokens": "^2.1.0",
-    "lit": "^3.2.0",
+    "lit": "^3.2.1",
     "prism-esm": "^1.29.0-fix.6",
     "tslib": "^2.7.0",
     "web-dev-server-plugin-lit-css": "^3.0.1"
   },
   "devDependencies": {
     "@11ty/eleventy": "^3.0.0",
-    "@11ty/eleventy-img": "^5.0.0",
+    "@11ty/eleventy-fetch": "^4.0.1",
     "@11ty/eleventy-plugin-rss": "^2.0.2",
     "@commitlint/cli": "^19.5.0",
     "@commitlint/config-conventional": "^19.5.0",
@@ -308,38 +312,39 @@
     "@lit/reactive-element": "^2.0.4",
     "@parse5/tools": "^0.5.0",
     "@patternfly/create-element": "^1.0.3",
-    "@patternfly/elements": "^4.0.1",
+    "@patternfly/elements": "^4.0.2",
     "@patternfly/eslint-config-elements": "^4.0.0",
     "@patternfly/eslint-plugin-elements": "^2.0.0",
     "@patternfly/icons": "^1.0.3",
-    "@patternfly/pfe-tools": "^4.0.0",
-    "@playwright/test": "^1.47.2",
+    "@patternfly/pfe-tools": "^4.0.1",
+    "@playwright/test": "^1.48.0",
     "@rollup/plugin-node-resolve": "^15.3.0",
-    "@rollup/plugin-typescript": "^11.1.6",
+    "@stylistic/eslint-plugin-js": "^2.9.0",
     "@stylistic/stylelint-config": "^2.0.0",
-    "@stylistic/stylelint-plugin": "^3.1.0",
-    "@types/mocha": "^10.0.8",
+    "@stylistic/stylelint-plugin": "^3.1.1",
+    "@types/js-yaml": "^4.0.9",
+    "@types/mocha": "^10.0.9",
     "@web/rollup-plugin-import-meta-assets": "^2.2.1",
     "@webcomponents/template-shadowroot": "^0.2.1",
     "async-csv": "^2.1.3",
     "autoprefixer": "^10.4.20",
     "cssnano": "^7.0.6",
     "custom-elements-manifest": "^2.1.0",
-    "element-internals-polyfill": "^1.3.11",
+    "element-internals-polyfill": "^1.3.12",
     "eleventy-plugin-dart-sass": "^1.0.3",
     "eleventy-plugin-helmet": "^0.2.2",
     "eleventy-plugin-nesting-toc": "^1.3.0",
     "es-module-shims": "^1.10.0",
-    "eslint": "^9.11.1",
+    "eslint": "^9.12.0",
     "fuse.js": "^7.0.0",
     "git-branch": "^2.0.1",
     "image-size": "^1.1.1",
     "leasot": "^14.4.0",
-    "lit-html": "^3.2.0",
+    "lit-html": "^3.2.1",
     "markdown-it-attrs": "^4.2.0",
     "markdown-it-footnote": "^4.0.0",
     "markdownlint-cli2": "^0.14.0",
-    "parse5": "^7.1.2",
+    "parse5": "^7.2.0",
     "patch-package": "^8.0.0",
     "playground-elements": "^0.19.1",
     "postcss-pxtorem": "^6.1.0",
@@ -350,7 +355,7 @@
     "tinycolor2": "^1.6.0",
     "ts-patch": "^3.2.1",
     "tsx": "^4.19.1",
-    "typescript": "^5.6.2",
+    "typescript": "^5.6.3",
     "typescript-transform-lit-css": "^2.0.0",
     "wireit": "^0.14.9"
   },
