@@ -72,7 +72,7 @@ export default class IconsPage {
         <ul class="icon-set">${this.#renderIcons('social', icons)}</ul>
       </section>
 
-      ${await this.renderFile('./docs/_includes/partials/component/feedback.11ty.cjs', ctx)}
+      ${await this.renderFile('./docs/_includes/partials/component/feedback.11ty.ts', ctx)}
     `;
   }
 
