@@ -44,9 +44,10 @@ export class IconResolveErrorEvent extends ErrorEvent {
 }
 
 /**
- * Icons represents general concepts and can support text as a decorative element.
- * The icon element, <rh-icon>, is a container that allows users to add icons of varying
- * dimensions in the same area without shifting surrounding content.
+ * Icons represents general concepts and can support text as a decorative
+ * element. The icon element, `<rh-icon>`, is a container that allows users to
+ * add icons of varying dimensions in the same area without shifting surrounding
+ * content.
  * @summary Decorative element which supports related content
  * @slot - Slotted content is used as a fallback in case the icon doesn't load
  * @fires load - Fired when an icon is loaded and rendered
