@@ -109,12 +109,8 @@ export class RhSwitch extends LitElement {
     const rtl = this.#dir.dir === 'rtl';
     const { on = 'light', reversed, checked } = this;
     const slots = html`
-      <slot class="message" name="message-on" ?hidden="${!this.checked}">
-        <span aria-hidden="true">${this.messageOn}</span>
-      </slot>
-      <slot class="message" name="message-off" ?hidden="${this.checked}">
-        <span aria-hidden="true">${this.messageOff}</span>
-      </slot>`;
+      <slot class="message" name="message-on" ?hidden="${!this.checked}"><span aria-hidden="true">${this.messageOn}</span></slot>
+      <slot class="message" name="message-off" ?hidden="${this.checked}"><span aria-hidden="true">${this.messageOff}</span></slot>`;
     return html`
       <div id="container"
            part="container"
